+++ lib/src/mime_shared.dart
@@ -1,16 +1,14 @@
 // Copyright (c) 2014, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
-library mime.shared;
-
-import 'dart:async';
 
 class MimeMultipartException implements Exception {
   final String message;
 
-  const MimeMultipartException([this.message = ""]);
+  const MimeMultipartException([this.message = '']);
 
-  String toString() => "MimeMultipartException: $message";
+  @override
+  String toString() => 'MimeMultipartException: $message';
 }
 
 /// A Mime Multipart class representing each part parsed by