+++ lib/src/mime_type.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library mime.mime_type;
-
 import 'default_extension_map.dart';
 import 'magic_number.dart';
 
@@ -22,8 +20,22 @@ int get defaultMagicNumbersMaxLength => _globalResolver.magicNumbersMaxLength;
 /// a file have been saved using the wrong file-name extension. If less than
 /// [defaultMagicNumbersMaxLength] bytes was provided, some magic-numbers won't
 /// be matched against.
-String lookupMimeType(String path, {List<int> headerBytes}) =>
+String? lookupMimeType(String path, {List<int>? headerBytes}) =>
     _globalResolver.lookup(path, headerBytes: headerBytes);
+
+/// Returns the extension for the given MIME type.
+///
+/// If there are multiple extensions for [mime], return the first occurrence in
+/// the map. If there are no extensions for [mime], return [mime].
+String extensionFromMime(String mime) {
+  mime = mime.toLowerCase();
+  for (final entry in defaultExtensionMap.entries) {
+    if (defaultExtensionMap[entry.key] == mime) {
+      return entry.key;
+    }
+  }
+  return mime;
+}
 
 /// MIME-type resolver class, used to customize the lookup of mime-types.
 class MimeTypeResolver {
@@ -40,7 +52,7 @@ class MimeTypeResolver {
   /// Create a new [MimeTypeResolver] containing the default scope.
   MimeTypeResolver()
       : _useDefault = true,
-        _magicNumbersMaxLength = DEFAULT_MAGIC_NUMBERS_MAX_LENGTH;
+        _magicNumbersMaxLength = initialMagicNumbersMaxLength;
 
   /// Get the maximum number of bytes required to match all magic numbers, when
   /// performing [lookup] with headerBytes present.
@@ -55,13 +67,13 @@ class MimeTypeResolver {
   /// though a file have been saved using the wrong file-name extension. If less
   /// than [magicNumbersMaxLength] bytes was provided, some magic-numbers won't
   /// be matched against.
-  String lookup(String path, {List<int> headerBytes}) {
-    String result;
+  String? lookup(String path, {List<int>? headerBytes}) {
+    String? result;
     if (headerBytes != null) {
       result = _matchMagic(headerBytes, _magicNumbers);
       if (result != null) return result;
       if (_useDefault) {
-        result = _matchMagic(headerBytes, DEFAULT_MAGIC_NUMBERS);
+        result = _matchMagic(headerBytes, initialMagicNumbers);
         if (result != null) return result;
       }
     }
@@ -85,7 +97,7 @@ class MimeTypeResolver {
   ///
   /// If [mask] is present,the [mask] is used to only perform matching on
   /// selective bits. The [mask] must have the same length as [bytes].
-  void addMagicNumber(List<int> bytes, String mimeType, {List<int> mask}) {
+  void addMagicNumber(List<int> bytes, String mimeType, {List<int>? mask}) {
     if (mask != null && bytes.length != mask.length) {
       throw ArgumentError('Bytes and mask are of different lengths');
     }
@@ -95,7 +107,7 @@ class MimeTypeResolver {
     _magicNumbers.add(MagicNumber(mimeType, bytes, mask: mask));
   }
 
-  static String _matchMagic(
+  static String? _matchMagic(
       List<int> headerBytes, List<MagicNumber> magicNumbers) {
     for (var mn in magicNumbers) {
       if (mn.matches(headerBytes)) return mn.mimeType;
@@ -104,7 +116,7 @@ class MimeTypeResolver {
   }
 
   static String _ext(String path) {
-    int index = path.lastIndexOf('.');
+    var index = path.lastIndexOf('.');
     if (index < 0 || index + 1 >= path.length) return path;
     return path.substring(index + 1).toLowerCase();
   }