+++ test/mime_multipart_transformer_test.dart
@@ -5,77 +5,90 @@
 import 'dart:async';
 import 'dart:math';
 
-import "package:test/test.dart";
-import "package:mime/mime.dart";
+import 'package:mime/mime.dart';
+import 'package:test/test.dart';
 
 void _writeInChunks(
     List<int> data, int chunkSize, StreamController<List<int>> controller) {
   if (chunkSize == -1) chunkSize = data.length;
 
-  for (int pos = 0; pos < data.length; pos += chunkSize) {
-    int remaining = data.length - pos;
-    int writeLength = min(chunkSize, remaining);
+  for (var pos = 0; pos < data.length; pos += chunkSize) {
+    var remaining = data.length - pos;
+    var writeLength = min(chunkSize, remaining);
     controller.add(data.sublist(pos, pos + writeLength));
   }
   controller.close();
 }
 
-enum TestMode { IMMEDIATE_LISTEN, DELAY_LISTEN, PAUSE_RESUME }
+enum TestMode { immediateListen, delayListen, pauseResume }
 
-void _runParseTest(String message, String boundary, TestMode mode,
-    [List<Map> expectedHeaders, List expectedParts, bool expectError = false]) {
+void _runParseTest(
+  String message,
+  String boundary,
+  TestMode mode, [
+  List<Map>? expectedHeaders,
+  List<String?>? expectedParts,
+  bool expectError = false,
+]) {
   Future testWrite(List<int> data, [int chunkSize = -1]) {
     var controller = StreamController<List<int>>(sync: true);
 
     var stream =
         controller.stream.transform(MimeMultipartTransformer(boundary));
-    int i = 0;
+    var i = 0;
     var completer = Completer();
     var futures = <Future>[];
     stream.listen((multipart) {
-      int part = i++;
+      var part = i++;
       if (expectedHeaders != null) {
         expect(multipart.headers, equals(expectedHeaders[part]));
       }
       switch (mode) {
-        case TestMode.IMMEDIATE_LISTEN:
-          futures.add(multipart
-              .fold([], (buffer, data) => buffer..addAll(data)).then((data) {
-            if (expectedParts[part] != null) {
-              expect(data, equals(expectedParts[part].codeUnits));
+        case TestMode.immediateListen:
+          futures.add(multipart.fold<List<int>>(
+              [], (buffer, data) => buffer..addAll(data)).then((data) {
+            if (expectedParts?[part] != null) {
+              expect(data, equals(expectedParts?[part]!.codeUnits));
             }
           }));
           break;
 
-        case TestMode.DELAY_LISTEN:
-          futures.add(Future(() {
-            return multipart
-                .fold([], (buffer, data) => buffer..addAll(data)).then((data) {
-              if (expectedParts[part] != null) {
-                expect(data, equals(expectedParts[part].codeUnits));
-              }
-            });
-          }));
+        case TestMode.delayListen:
+          futures.add(
+            Future(
+              () => multipart.fold<List<int>>(
+                [],
+                (buffer, data) => buffer..addAll(data),
+              ).then(
+                (data) {
+                  if (expectedParts?[part] != null) {
+                    expect(data, equals(expectedParts?[part]!.codeUnits));
+                  }
+                },
+              ),
+            ),
+          );
           break;
 
-        case TestMode.PAUSE_RESUME:
+        case TestMode.pauseResume:
           var completer = Completer();
           futures.add(completer.future);
           var buffer = [];
-          var subscription;
+          late StreamSubscription subscription;
           subscription = multipart.listen((data) {
             buffer.addAll(data);
             subscription.pause();
             Future(() => subscription.resume());
           }, onDone: () {
-            if (expectedParts[part] != null) {
-              expect(buffer, equals(expectedParts[part].codeUnits));
+            if (expectedParts?[part] != null) {
+              expect(buffer, equals(expectedParts?[part]!.codeUnits));
             }
             completer.complete();
           });
+          addTearDown(subscription.cancel);
           break;
       }
-    }, onError: (error) {
+    }, onError: (Object error) {
       if (!expectError) throw error;
     }, onDone: () {
       if (expectedParts != null) {
@@ -100,11 +113,13 @@ void _runParseTest(String message, String boundary, TestMode mode,
       if (expectedHeaders != null) {
         expect(multipart.headers, equals(expectedHeaders[0]));
       }
-      return (multipart.fold([], (b, d) => b..addAll(d)).then((data) {
-        if (expectedParts != null && expectedParts[0] != null) {
-          expect(data, equals(expectedParts[0].codeUnits));
-        }
-      }));
+      return multipart.fold<List<int>>([], (b, d) => b..addAll(d)).then(
+        (data) {
+          if (expectedParts != null && expectedParts[0] != null) {
+            expect(data, equals(expectedParts[0]!.codeUnits));
+          }
+        },
+      );
     }).then((_) {
       completer.complete();
     });
@@ -120,11 +135,11 @@ void _runParseTest(String message, String boundary, TestMode mode,
     var controller = StreamController<List<int>>(sync: true);
     var stream =
         controller.stream.transform(MimeMultipartTransformer(boundary));
-    var subscription;
-    int i = 0;
+    late StreamSubscription subscription;
+    var i = 0;
     var futures = <Future>[];
     subscription = stream.listen((multipart) {
-      int partIndex = i;
+      var partIndex = i;
 
       if (partIndex >= parts) {
         throw StateError('Expected no more parts, but got one.');
@@ -133,11 +148,12 @@ void _runParseTest(String message, String boundary, TestMode mode,
       if (expectedHeaders != null) {
         expect(multipart.headers, equals(expectedHeaders[partIndex]));
       }
-      futures.add((multipart.fold([], (b, d) => b..addAll(d)).then((data) {
+      futures.add(
+          multipart.fold<List<int>>([], (b, d) => b..addAll(d)).then((data) {
         if (expectedParts != null && expectedParts[partIndex] != null) {
-          expect(data, equals(expectedParts[partIndex].codeUnits));
+          expect(data, equals(expectedParts[partIndex]!.codeUnits));
         }
-      })));
+      }));
 
       if (partIndex == (parts - 1)) {
         subscription.cancel();
@@ -153,7 +169,7 @@ void _runParseTest(String message, String boundary, TestMode mode,
 
   // Test parsing the data three times delivering the data in
   // different chunks.
-  List<int> data = message.codeUnits;
+  var data = message.codeUnits;
   test('test', () {
     expect(
         Future.wait([
@@ -165,7 +181,7 @@ void _runParseTest(String message, String boundary, TestMode mode,
         completes);
   });
 
-  if (expectedParts.isNotEmpty) {
+  if (expectedParts!.isNotEmpty) {
     test('test-first-part-only', () {
       expect(
           Future.wait([
@@ -178,7 +194,7 @@ void _runParseTest(String message, String boundary, TestMode mode,
     });
 
     test('test-n-parts-only', () {
-      int numPartsExpected = expectedParts.length - 1;
+      var numPartsExpected = expectedParts.length - 1;
       if (numPartsExpected == 0) numPartsExpected = 1;
 
       expect(
@@ -194,22 +210,24 @@ void _runParseTest(String message, String boundary, TestMode mode,
 }
 
 void _testParse(String message, String boundary,
-    [List<Map> expectedHeaders, List expectedParts, bool expectError = false]) {
-  _runParseTest(message, boundary, TestMode.IMMEDIATE_LISTEN, expectedHeaders,
+    [List<Map>? expectedHeaders,
+    List<String?>? expectedParts,
+    bool expectError = false]) {
+  _runParseTest(message, boundary, TestMode.immediateListen, expectedHeaders,
       expectedParts, expectError);
-  _runParseTest(message, boundary, TestMode.DELAY_LISTEN, expectedHeaders,
+  _runParseTest(message, boundary, TestMode.delayListen, expectedHeaders,
       expectedParts, expectError);
-  _runParseTest(message, boundary, TestMode.PAUSE_RESUME, expectedHeaders,
+  _runParseTest(message, boundary, TestMode.pauseResume, expectedHeaders,
       expectedParts, expectError);
 }
 
 void _testParseValid() {
   // Empty message from Chrome form post.
   var message = '------WebKitFormBoundaryU3FBruSkJKG0Yor1--\r\n';
-  _testParse(message, "----WebKitFormBoundaryU3FBruSkJKG0Yor1", [], []);
+  _testParse(message, '----WebKitFormBoundaryU3FBruSkJKG0Yor1', [], []);
 
   // Sample from Wikipedia.
-  message = """
+  message = '''
 This is a message with multiple parts in MIME format.\r
 --frontier\r
 Content-Type: text/plain\r
@@ -221,40 +239,40 @@ Content-Transfer-Encoding: base64\r
 \r
 PGh0bWw+CiAgPGhlYWQ+CiAgPC9oZWFkPgogIDxib2R5PgogICAgPHA+VGhpcyBpcyB0aGUg
 Ym9keSBvZiB0aGUgbWVzc2FnZS48L3A+CiAgPC9ib2R5Pgo8L2h0bWw+Cg=\r
---frontier--\r\n""";
-  var headers1 = <String, String>{"content-type": "text/plain"};
+--frontier--\r\n''';
+  var headers1 = <String, String>{'content-type': 'text/plain'};
   var headers2 = <String, String>{
-    "content-type": "application/octet-stream",
-    "content-transfer-encoding": "base64"
+    'content-type': 'application/octet-stream',
+    'content-transfer-encoding': 'base64'
   };
-  var body1 = "This is the body of the message.";
-  var body2 = """
+  var body1 = 'This is the body of the message.';
+  var body2 = '''
 PGh0bWw+CiAgPGhlYWQ+CiAgPC9oZWFkPgogIDxib2R5PgogICAgPHA+VGhpcyBpcyB0aGUg
-Ym9keSBvZiB0aGUgbWVzc2FnZS48L3A+CiAgPC9ib2R5Pgo8L2h0bWw+Cg=""";
-  _testParse(message, "frontier", [headers1, headers2], [body1, body2]);
+Ym9keSBvZiB0aGUgbWVzc2FnZS48L3A+CiAgPC9ib2R5Pgo8L2h0bWw+Cg=''';
+  _testParse(message, 'frontier', [headers1, headers2], [body1, body2]);
 
   // Sample from HTML 4.01 Specification.
-  message = """
+  message = '''
 \r\n--AaB03x\r
-Content-Disposition: form-data; name=\"submit-name\"\r
+Content-Disposition: form-data; name="submit-name"\r
 \r
 Larry\r
 --AaB03x\r
-Content-Disposition: form-data; name=\"files\"; filename=\"file1.txt\"\r
+Content-Disposition: form-data; name="files"; filename="file1.txt"\r
 Content-Type: text/plain\r
 \r
 ... contents of file1.txt ...\r
---AaB03x--\r\n""";
+--AaB03x--\r\n''';
   headers1 = <String, String>{
-    "content-disposition": "form-data; name=\"submit-name\""
+    'content-disposition': 'form-data; name="submit-name"'
   };
   headers2 = <String, String>{
-    "content-type": "text/plain",
-    "content-disposition": "form-data; name=\"files\"; filename=\"file1.txt\""
+    'content-type': 'text/plain',
+    'content-disposition': 'form-data; name="files"; filename="file1.txt"'
   };
-  body1 = "Larry";
-  body2 = "... contents of file1.txt ...";
-  _testParse(message, "AaB03x", [headers1, headers2], [body1, body2]);
+  body1 = 'Larry';
+  body2 = '... contents of file1.txt ...';
+  _testParse(message, 'AaB03x', [headers1, headers2], [body1, body2]);
 
   // Longer form from submitting the following from Chrome.
   //
@@ -262,104 +280,104 @@ Content-Type: text/plain\r
   // <body>
   // <FORM action="http://127.0.0.1:1234/"
   //     enctype="multipart/form-data"
-  //     method="post">
+  //     method='post'>
   //  <P>
-  //  Text: <INPUT type="text" name="text_input">
-  //  Password: <INPUT type="password" name="password_input">
-  //  Checkbox: <INPUT type="checkbox" name="checkbox_input">
-  //  Radio: <INPUT type="radio" name="radio_input">
-  //  Send <INPUT type="submit">
+  //  Text: <INPUT type='text' name='text_input'>
+  //  Password: <INPUT type='password' name='password_input'>
+  //  Checkbox: <INPUT type='checkbox' name='checkbox_input'>
+  //  Radio: <INPUT type='radio' name='radio_input'>
+  //  Send <INPUT type='submit'>
   //  </P>
   // </FORM>
   // </body>
   // </html>
 
-  message = """
+  message = '''
 \r\n------WebKitFormBoundaryQ3cgYAmGRF8yOeYB\r
-Content-Disposition: form-data; name=\"text_input\"\r
+Content-Disposition: form-data; name="text_input"\r
 \r
 text\r
 ------WebKitFormBoundaryQ3cgYAmGRF8yOeYB\r
-Content-Disposition: form-data; name=\"password_input\"\r
+Content-Disposition: form-data; name="password_input"\r
 \r
 password\r
 ------WebKitFormBoundaryQ3cgYAmGRF8yOeYB\r
-Content-Disposition: form-data; name=\"checkbox_input\"\r
+Content-Disposition: form-data; name="checkbox_input"\r
 \r
 on\r
 ------WebKitFormBoundaryQ3cgYAmGRF8yOeYB\r
-Content-Disposition: form-data; name=\"radio_input\"\r
+Content-Disposition: form-data; name="radio_input"\r
 \r
 on\r
-------WebKitFormBoundaryQ3cgYAmGRF8yOeYB--\r\n""";
+------WebKitFormBoundaryQ3cgYAmGRF8yOeYB--\r\n''';
   headers1 = <String, String>{
-    "content-disposition": "form-data; name=\"text_input\""
+    'content-disposition': 'form-data; name="text_input"'
   };
   headers2 = <String, String>{
-    "content-disposition": "form-data; name=\"password_input\""
+    'content-disposition': 'form-data; name="password_input"'
   };
   var headers3 = <String, String>{
-    "content-disposition": "form-data; name=\"checkbox_input\""
+    'content-disposition': 'form-data; name="checkbox_input"'
   };
   var headers4 = <String, String>{
-    "content-disposition": "form-data; name=\"radio_input\""
+    'content-disposition': 'form-data; name="radio_input"'
   };
-  body1 = "text";
-  body2 = "password";
-  var body3 = "on";
-  var body4 = "on";
-  _testParse(message, "----WebKitFormBoundaryQ3cgYAmGRF8yOeYB",
+  body1 = 'text';
+  body2 = 'password';
+  var body3 = 'on';
+  var body4 = 'on';
+  _testParse(message, '----WebKitFormBoundaryQ3cgYAmGRF8yOeYB',
       [headers1, headers2, headers3, headers4], [body1, body2, body3, body4]);
 
   // Same form from Firefox.
-  message = """
+  message = '''
 \r\n-----------------------------52284550912143824192005403738\r
-Content-Disposition: form-data; name=\"text_input\"\r
+Content-Disposition: form-data; name="text_input"\r
 \r
 text\r
 -----------------------------52284550912143824192005403738\r
-Content-Disposition: form-data; name=\"password_input\"\r
+Content-Disposition: form-data; name="password_input"\r
 \r
 password\r
 -----------------------------52284550912143824192005403738\r
-Content-Disposition: form-data; name=\"checkbox_input\"\r
+Content-Disposition: form-data; name="checkbox_input"\r
 \r
 on\r
 -----------------------------52284550912143824192005403738\r
-Content-Disposition: form-data; name=\"radio_input\"\r
+Content-Disposition: form-data; name="radio_input"\r
 \r
 on\r
------------------------------52284550912143824192005403738--\r\n""";
+-----------------------------52284550912143824192005403738--\r\n''';
   _testParse(
       message,
-      "---------------------------52284550912143824192005403738",
+      '---------------------------52284550912143824192005403738',
       [headers1, headers2, headers3, headers4],
       [body1, body2, body3, body4]);
 
   // And Internet Explorer
-  message = """
+  message = '''
 \r\n-----------------------------7dc8f38c60326\r
-Content-Disposition: form-data; name=\"text_input\"\r
+Content-Disposition: form-data; name="text_input"\r
 \r
 text\r
 -----------------------------7dc8f38c60326\r
-Content-Disposition: form-data; name=\"password_input\"\r
+Content-Disposition: form-data; name="password_input"\r
 \r
 password\r
 -----------------------------7dc8f38c60326\r
-Content-Disposition: form-data; name=\"checkbox_input\"\r
+Content-Disposition: form-data; name="checkbox_input"\r
 \r
 on\r
 -----------------------------7dc8f38c60326\r
-Content-Disposition: form-data; name=\"radio_input\"\r
+Content-Disposition: form-data; name="radio_input"\r
 \r
 on\r
------------------------------7dc8f38c60326--\r\n""";
-  _testParse(message, "---------------------------7dc8f38c60326",
+-----------------------------7dc8f38c60326--\r\n''';
+  _testParse(message, '---------------------------7dc8f38c60326',
       [headers1, headers2, headers3, headers4], [body1, body2, body3, body4]);
 
   // Test boundary prefix inside prefix and content.
-  message = """
+  message = '''
 -\r
 --\r
 --b\r
@@ -393,9 +411,9 @@ Content-Type: text/plain\r
 --b\r\r\r\r
 --\r\r\r
 -\r\r
---boundary--\r\n""";
-  var headers = <String, String>{"content-type": "text/plain"};
-  body1 = """
+--boundary--\r\n''';
+  var headers = <String, String>{'content-type': 'text/plain'};
+  body1 = '''
 -\r
 --\r
 --b\r
@@ -404,8 +422,8 @@ Content-Type: text/plain\r
 --boun\r
 --bound\r\r
 --bounda\r\r\r
---boundar\r\r\r""";
-  body2 = """
+--boundar\r\r\r''';
+  body2 = '''
 --boundar\r
 --bounda\r
 --bound\r
@@ -414,11 +432,11 @@ Content-Type: text/plain\r
 --bo\r
 --b\r\r\r\r
 --\r\r\r
--\r""";
-  _testParse(message, "boundary", [headers, headers], [body1, body2]);
+-\r''';
+  _testParse(message, 'boundary', [headers, headers], [body1, body2]);
 
   // Without initial CRLF.
-  message = """
+  message = '''
 --xxx\r
 \r
 \r
@@ -427,13 +445,13 @@ Body 1\r
 \r
 \r
 Body2\r
---xxx--\r\n""";
-  _testParse(message, "xxx", null, ["\r\nBody 1", "\r\nBody2"]);
+--xxx--\r\n''';
+  _testParse(message, 'xxx', null, ['\r\nBody 1', '\r\nBody2']);
 }
 
 void _testParseInvalid() {
   // Missing end boundary.
-  var message = """
+  var message = '''
 \r
 --xxx\r
 \r
@@ -443,8 +461,8 @@ Body 1\r
 \r
 \r
 Body2\r
---xxx\r\n""";
-  _testParse(message, "xxx", null, [null, null], true);
+--xxx\r\n''';
+  _testParse(message, 'xxx', null, [null, null], true);
 }
 
 void main() {