+++ test/mime_type_test.dart
@@ -4,13 +4,13 @@
 
 import 'dart:math' as math;
 
-import 'package:test/test.dart';
 import 'package:mime/mime.dart';
 import 'package:mime/src/magic_number.dart';
+import 'package:test/test.dart';
 
-void _expectMimeType(String path, String expectedMimeType,
-    {List<int> headerBytes, MimeTypeResolver resolver}) {
-  String mimeType;
+void _expectMimeType(String path, String? expectedMimeType,
+    {List<int>? headerBytes, MimeTypeResolver? resolver}) {
+  String? mimeType;
   if (resolver == null) {
     mimeType = lookupMimeType(path, headerBytes: headerBytes);
   } else {
@@ -40,6 +40,13 @@ void main() {
       _expectMimeType('file.pdf', 'application/pdf');
       _expectMimeType('file.tiff', 'image/tiff');
       _expectMimeType('file.tif', 'image/tiff');
+      _expectMimeType('file.webp', 'image/webp');
+      _expectMimeType('file.mp3', 'audio/mpeg');
+      _expectMimeType('file.aac', 'audio/aac');
+      _expectMimeType('file.ogg', 'audio/ogg');
+      _expectMimeType('file.aiff', 'audio/x-aiff');
+      _expectMimeType('file.m4a', 'audio/mp4');
+      _expectMimeType('file.toml', 'application/toml');
     });
 
     test('unknown-mime-type', () {
@@ -87,6 +94,56 @@ void main() {
         0x70,
         0x35
       ]);
+      _expectMimeType('file', 'image/webp', headerBytes: [
+        0x52,
+        0x49,
+        0x46,
+        0x46,
+        0xE2,
+        0x4A,
+        0x01,
+        0x00,
+        0x57,
+        0x45,
+        0x42,
+        0x50
+      ]);
+      _expectMimeType('file', 'audio/mpeg',
+          headerBytes: [0x49, 0x44, 0x33, 0x0D, 0x0A, 0x1A, 0x0A]);
+      _expectMimeType('file', 'audio/aac',
+          headerBytes: [0xFF, 0xF1, 0x0D, 0x0A, 0x1A, 0x0A]);
+      _expectMimeType('file', 'audio/ogg',
+          headerBytes: [0x4F, 0x70, 0x75, 0x0D, 0x0A, 0x1A, 0x0A]);
+      _expectMimeType('file', 'audio/x-aiff', headerBytes: [
+        0x46,
+        0x4F,
+        0x52,
+        0x4D,
+        0x04,
+        0x0B,
+        0xEF,
+        0xF4,
+        0x41,
+        0x49,
+        0x46,
+        0x46
+      ]);
+      _expectMimeType('file', 'audio/x-flac',
+          headerBytes: [0x66, 0x4C, 0x61, 0x43]);
+      _expectMimeType('file', 'audio/x-wav', headerBytes: [
+        0x52,
+        0x49,
+        0x46,
+        0x46,
+        0xA6,
+        0x4E,
+        0x70,
+        0x03,
+        0x57,
+        0x41,
+        0x56,
+        0x45
+      ]);
     });
   });
 
@@ -117,10 +174,29 @@ void main() {
   });
 
   test('default magic number', () {
-    var actualMaxBytes = DEFAULT_MAGIC_NUMBERS.fold(0, (num previous, magic) {
-      return math.max(previous, magic.numbers.length);
+    var actualMaxBytes = initialMagicNumbers.fold<int>(
+      0,
+      (previous, magic) => math.max(previous, magic.numbers.length),
+    );
+
+    expect(initialMagicNumbersMaxLength, actualMaxBytes);
+  });
+
+  group('extensionFromMime', () {
+    test('returns match for mime with single extension', () {
+      expect(extensionFromMime('application/json'), equals('json'));
+      expect(extensionFromMime('application/java-archive'), equals('jar'));
     });
 
-    expect(defaultMagicNumbersMaxLength, actualMaxBytes);
+    test('returns first match for mime with multiple extensions', () {
+      expect(extensionFromMime('text/html'), equals('htm'));
+      expect(extensionFromMime('application/x-cbr'), equals('cb7'));
+    });
+
+    test('returns inputted string for unrecognized mime', () {
+      expect(
+          extensionFromMime('unrecognized_mime'), equals('unrecognized_mime'));
+      expect(extensionFromMime('i/am/not/a/mime'), equals('i/am/not/a/mime'));
+    });
   });
 }